fixes kubectl replace that name is not a required argument
