--- e2e/src/test/groovy/io/kubernetes/client/e2e/kubectl/KubectlLabelTest.groovy
@@ -1,18 +1,0 @@
-/*
-Copyright 2020 The Kubernetes Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package io.kubernetes.client.e2e.kubectl
-
-import spock.lang.Specification
-
-class KubectlLabelTest extends Specification {
-}
